asdf

+++ LawyersApp/LawyersApp/Controllers/CourtsController.cs
@@ -7,14 +7,20 @@ using System.Web;
 using System.Web.Mvc;
 using Kendo.Mvc.UI;
 using Kendo.Mvc.Extensions;
+using LawyersApp.Models.Foreignkey;
+using LawyersApp.Models.Attributes;
+
 
 namespace LawyersApp.Controllers
 {
     public class CourtsController : Controller
     {
         // GET: Courts
+        [SessionTimeout]
+        [HttpGet]
         public ActionResult Courts()
         {
+            PopulateIssuesType();
             return View();
         }
 
@@ -23,6 +29,22 @@ namespace LawyersApp.Controllers
         {
             CourtsService = new CourtsService(new LawyersEntities());
         }
+
+        private void PopulateIssuesType()
+        {
+            var dataContext = new LawyersEntities();
+            var IssuesType = dataContext.IssuesType_Table
+                        .Select(c => new IssuesTypeForeingKey
+                        {
+                            IssuesTypeID = c.IssuesTypeID,
+                            IssuesTypename = c.IssuesTypename
+                        })
+                        .OrderBy(e => e.IssuesTypeID);
+
+            ViewData["IssuesType"] = IssuesType;
+            ViewData["defaultIssuesType"] = IssuesType.First();
+        }
+
         public ActionResult Courts_Read([DataSourceRequest] DataSourceRequest request)
         {
             return Json(CourtsService.Read().ToDataSourceResult(request));
@@ -56,13 +78,32 @@ namespace LawyersApp.Controllers
 
         public ActionResult Courts_Destroy([DataSourceRequest] DataSourceRequest request, CourtsViewModel db)
         {
-            if (db != null)
+            try
             {
-                CourtsService.Destroy(db);
-            }
+                if (ModelState.IsValid)
+                {
+                    if (db != null)
+                    {
+                        CourtsService.Destroy(db);
+                    }
 
+                }
+            }
+            catch (Exception)
+            {
+                ModelState.AddModelError("خطأ", "لا يمكن الحذف");
+                return Json(new[] { db }.ToDataSourceResult(request, ModelState));
+
+            }
             return Json(new[] { db }.ToDataSourceResult(request, ModelState));
+
         }
+
+        public JsonResult GetIssuesType()
+        {
+            return Json(CourtsService.GetIssuesType(), JsonRequestBehavior.AllowGet);
+        }
+
 
     }
 }
